Task #3: Refactor code
