+++ src/components/ChatBox/ChatBox.tsx
@@ -3,8 +3,6 @@ import { ChatList } from './ChatList'
 import { ChatTools, SearchChat } from './ChatTools'
 
 export const ChatBox = () => {
-  console.log('chatbox rendered')
-
   return (
     <div className="flex flex-col h-full py-2 max-w-[20rem] max-sm:hidden">
       <ChatTools />
